backend/development_tracker/requirements.txt
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf_spectacular==0.26.5
gunicorn==21.2.0
psycopg==3.1.9
psycopg-binary==3.1.9
python-dotenv==1.0.0
